# [minor] Reduce code without loss of performance