@lpinca 2017-12-04
Seems good, how many bytes are saved after minification? Can you also post a standalone benchmark of the original and proposed way to copy arguments on older Node.js versions (e.g. 4 and 6)?

Thank you.