@hongmaoxiao 2017-12-04
build file:<br>
original:
```
❯ ll umd/eventemitter3.min.js 
-rw-rw-r-- 1 mao mao 3518 12月  4 20:51 umd/eventemitter3.min.js
```

new:
```
❯ ll umd/eventemitter3.min.js
-rw-rw-r-- 1 mao mao 3498 12月  4 20:49 umd/eventemitter3.min.js
```