@hongmaoxiao 2017-12-04
**env**<br>
```
node: 4.6.2
npm: 2.15.11
```
**original**
```
~/learn/ljs/eventemitter3 master* 2m 14s
❯ npm run benchmark

> eventemitter3@3.0.0 benchmark /home/mao/learn/ljs/eventemitter3
> find benchmarks/run -name '*.js' -exec benchmarks/start.sh {} \;

Starting benchmark listeners.js

EventEmitter1 x 10,864,678 ops/sec ±1.93% (86 runs sampled)
EventEmitter3@0.1.6 x 4,159,596 ops/sec ±4.56% (80 runs sampled)
EventEmitter3(master) x 9,194,100 ops/sec ±3.91% (82 runs sampled)
fastemitter x 5,590,926 ops/sec ±3.54% (80 runs sampled)
Fastest is EventEmitter1

Starting benchmark init.js

EventEmitter1 x 33,379,683 ops/sec ±3.18% (86 runs sampled)
EventEmitter2 x 31,378,051 ops/sec ±3.81% (78 runs sampled)
EventEmitter3@0.1.6 x 73,284,120 ops/sec ±3.02% (82 runs sampled)
EventEmitter3(master) x 57,668,859 ops/sec ±3.11% (86 runs sampled)
Drip x 68,116,413 ops/sec ±4.66% (76 runs sampled)
fastemitter x 15,697,396 ops/sec ±3.08% (84 runs sampled)
event-emitter x 9,541,304 ops/sec ±2.82% (84 runs sampled)
contra/emitter x 908,929 ops/sec ±3.15% (65 runs sampled)
Fastest is EventEmitter3@0.1.6

Starting benchmark once.js

EventEmitter1 x 3,263,313 ops/sec ±3.55% (81 runs sampled)
EventEmitter2 x 1,281,131 ops/sec ±4.37% (75 runs sampled)
EventEmitter3@0.1.6 x 1,853,770 ops/sec ±2.61% (81 runs sampled)
EventEmitter3(master) x 7,569,117 ops/sec ±4.10% (82 runs sampled)
Drip x 6,097,429 ops/sec ±1.89% (82 runs sampled)
fastemitter x 5,061,037 ops/sec ±2.02% (82 runs sampled)
event-emitter x 1,215,815 ops/sec ±3.19% (81 runs sampled)
contra/emitter x 502,581 ops/sec ±3.45% (87 runs sampled)
Fastest is EventEmitter3(master)

Starting benchmark emit.js

EventEmitter1 x 9,022,124 ops/sec ±3.07% (83 runs sampled)
EventEmitter2 x 5,741,463 ops/sec ±3.29% (84 runs sampled)
EventEmitter3@0.1.6 x 10,156,406 ops/sec ±2.47% (88 runs sampled)
EventEmitter3(master) x 9,683,323 ops/sec ±4.04% (81 runs sampled)
Drip x 5,895,005 ops/sec ±3.97% (78 runs sampled)
fastemitter x 4,905,202 ops/sec ±3.41% (83 runs sampled)
event-emitter x 2,457,011 ops/sec ±1.48% (80 runs sampled)
contra/emitter x 104,733 ops/sec ±1.49% (83 runs sampled)
Fastest is EventEmitter3@0.1.6,EventEmitter3(master)

Starting benchmark hundreds.js

EventEmitter1 x 375,776 ops/sec ±2.69% (86 runs sampled)
EventEmitter2 x 256,028 ops/sec ±2.17% (88 runs sampled)
EventEmitter3@0.1.6 x 312,508 ops/sec ±4.19% (78 runs sampled)
EventEmitter3(master) x 342,425 ops/sec ±3.55% (89 runs sampled)
Drip x 447,544 ops/sec ±3.14% (85 runs sampled)
fastemitter x 422,861 ops/sec ±4.31% (74 runs sampled)
event-emitter x 146,425 ops/sec ±2.80% (87 runs sampled)
contra/emitter x 40,524 ops/sec ±3.96% (83 runs sampled)
Fastest is Drip,fastemitter

Starting benchmark add-remove.js

EventEmitter1 x 5,545,580 ops/sec ±4.67% (73 runs sampled)
EventEmitter2 x 1,822,583 ops/sec ±3.68% (77 runs sampled)
EventEmitter3@0.1.6 x 2,041,607 ops/sec ±3.47% (81 runs sampled)
EventEmitter3(master) x 7,946,913 ops/sec ±5.62% (77 runs sampled)
Drip x 24,616,436 ops/sec ±4.22% (82 runs sampled)
fastemitter x 17,661,798 ops/sec ±5.00% (79 runs sampled)
event-emitter x 1,807,188 ops/sec ±3.63% (82 runs sampled)
contra/emitter x 7,082,249 ops/sec ±2.60% (82 runs sampled)
Fastest is Drip

Starting benchmark remove-emit.js

EventEmitter1 x 5,116,395 ops/sec ±4.22% (75 runs sampled)
EventEmitter2 x 3,670,787 ops/sec ±4.55% (77 runs sampled)
EventEmitter3@0.1.6 x 3,846,422 ops/sec ±4.14% (79 runs sampled)
EventEmitter3(master) x 4,062,447 ops/sec ±3.07% (80 runs sampled)
Drip x 5,957,976 ops/sec ±2.90% (71 runs sampled)
event-emitter x 1,796,758 ops/sec ±3.64% (83 runs sampled)
contra/emitter x 109,918 ops/sec ±3.02% (84 runs sampled)
Fastest is Drip

Starting benchmark emit-multiple-listeners.js

EventEmitter1 x 3,418,490 ops/sec ±3.74% (76 runs sampled)
EventEmitter2 x 1,170,897 ops/sec ±1.96% (85 runs sampled)
EventEmitter3@0.1.6 x 2,533,289 ops/sec ±1.85% (84 runs sampled)
EventEmitter3(master) x 4,700,625 ops/sec ±2.18% (87 runs sampled)
fastemitter x 2,540,936 ops/sec ±2.73% (87 runs sampled)
event-emitter x 657,388 ops/sec ±3.04% (83 runs sampled)
contra/emitter x 99,472 ops/sec ±2.65% (87 runs sampled)
Fastest is EventEmitter3(master)

Starting benchmark context.js

EventEmitter1 x 938,388 ops/sec ±3.86% (83 runs sampled)
EventEmitter2 x 881,666 ops/sec ±3.35% (84 runs sampled)
EventEmitter3@0.1.6 x 9,214,864 ops/sec ±4.29% (78 runs sampled)
EventEmitter3(master) x 9,302,721 ops/sec ±4.03% (85 runs sampled)
Drip x 865,679 ops/sec ±4.23% (74 runs sampled)
fastemitter x 851,660 ops/sec ±2.94% (89 runs sampled)
event-emitter x 679,391 ops/sec ±4.71% (75 runs sampled)
contra/emitter x 94,118 ops/sec ±3.48% (84 runs sampled)
Fastest is EventEmitter3(master)
```
**new**
```
~/fork/eventemitter3 master*
❯ node -v
v4.6.2

~/fork/eventemitter3 master*
❯ npm run benchmark

> eventemitter3@3.0.0 benchmark /home/mao/fork/eventemitter3
> find benchmarks/run -name '*.js' -exec benchmarks/start.sh {} \;

Starting benchmark listeners.js

EventEmitter1 x 10,321,153 ops/sec ±2.40% (84 runs sampled)
EventEmitter3@0.1.6 x 4,229,339 ops/sec ±2.72% (89 runs sampled)
EventEmitter3(master) x 9,709,144 ops/sec ±2.14% (84 runs sampled)
fastemitter x 5,995,888 ops/sec ±1.62% (80 runs sampled)
Fastest is EventEmitter1

Starting benchmark init.js

EventEmitter1 x 36,100,829 ops/sec ±1.28% (90 runs sampled)
EventEmitter2 x 31,972,441 ops/sec ±2.48% (88 runs sampled)
EventEmitter3@0.1.6 x 76,100,136 ops/sec ±2.51% (82 runs sampled)
EventEmitter3(master) x 54,348,070 ops/sec ±1.94% (81 runs sampled)
Drip x 70,040,266 ops/sec ±2.40% (80 runs sampled)
fastemitter x 14,536,025 ops/sec ±2.95% (82 runs sampled)
event-emitter x 9,396,871 ops/sec ±2.52% (86 runs sampled)
contra/emitter x 1,317,935 ops/sec ±1.99% (77 runs sampled)
Fastest is EventEmitter3@0.1.6

Starting benchmark once.js

EventEmitter1 x 3,617,744 ops/sec ±1.77% (91 runs sampled)
EventEmitter2 x 1,422,960 ops/sec ±1.82% (83 runs sampled)
EventEmitter3@0.1.6 x 1,867,505 ops/sec ±2.90% (79 runs sampled)
EventEmitter3(master) x 7,354,929 ops/sec ±4.09% (82 runs sampled)
Drip x 6,410,944 ops/sec ±2.55% (86 runs sampled)
fastemitter x 5,022,907 ops/sec ±3.87% (75 runs sampled)
event-emitter x 1,285,471 ops/sec ±2.59% (72 runs sampled)
contra/emitter x 511,326 ops/sec ±2.37% (82 runs sampled)
Fastest is EventEmitter3(master)

Starting benchmark emit.js

EventEmitter1 x 9,255,834 ops/sec ±1.74% (88 runs sampled)
EventEmitter2 x 5,910,713 ops/sec ±1.11% (84 runs sampled)
EventEmitter3@0.1.6 x 9,868,930 ops/sec ±2.07% (81 runs sampled)
EventEmitter3(master) x 10,132,968 ops/sec ±2.47% (81 runs sampled)
Drip x 5,974,232 ops/sec ±2.38% (83 runs sampled)
fastemitter x 4,881,050 ops/sec ±1.19% (86 runs sampled)
event-emitter x 2,331,075 ops/sec ±0.97% (90 runs sampled)
contra/emitter x 105,158 ops/sec ±0.88% (92 runs sampled)
Fastest is EventEmitter3(master)

Starting benchmark hundreds.js

EventEmitter1 x 365,412 ops/sec ±1.82% (86 runs sampled)
EventEmitter2 x 256,301 ops/sec ±1.61% (81 runs sampled)
EventEmitter3@0.1.6 x 314,293 ops/sec ±3.63% (86 runs sampled)
EventEmitter3(master) x 343,120 ops/sec ±2.31% (87 runs sampled)
Drip x 438,516 ops/sec ±3.26% (78 runs sampled)
fastemitter x 443,859 ops/sec ±3.54% (88 runs sampled)
event-emitter x 146,619 ops/sec ±3.47% (82 runs sampled)
contra/emitter x 43,957 ops/sec ±2.26% (84 runs sampled)
Fastest is fastemitter

Starting benchmark add-remove.js

EventEmitter1 x 5,852,860 ops/sec ±4.27% (78 runs sampled)
EventEmitter2 x 1,873,556 ops/sec ±3.05% (87 runs sampled)
EventEmitter3@0.1.6 x 2,099,729 ops/sec ±2.91% (84 runs sampled)
EventEmitter3(master) x 8,843,894 ops/sec ±2.28% (85 runs sampled)
Drip x 28,312,703 ops/sec ±1.77% (86 runs sampled)
fastemitter x 18,908,371 ops/sec ±1.65% (91 runs sampled)
event-emitter x 1,805,255 ops/sec ±1.01% (71 runs sampled)
contra/emitter x 7,236,877 ops/sec ±1.89% (88 runs sampled)
Fastest is Drip

Starting benchmark remove-emit.js

EventEmitter1 x 5,418,895 ops/sec ±1.54% (92 runs sampled)
EventEmitter2 x 4,076,245 ops/sec ±1.65% (86 runs sampled)
EventEmitter3@0.1.6 x 4,289,941 ops/sec ±1.79% (89 runs sampled)
EventEmitter3(master) x 4,186,586 ops/sec ±1.01% (86 runs sampled)
Drip x 5,933,888 ops/sec ±1.66% (76 runs sampled)
event-emitter x 1,831,667 ops/sec ±1.67% (81 runs sampled)
contra/emitter x 105,652 ops/sec ±2.61% (87 runs sampled)
Fastest is Drip

Starting benchmark emit-multiple-listeners.js

EventEmitter1 x 3,446,988 ops/sec ±2.34% (90 runs sampled)
EventEmitter2 x 1,144,086 ops/sec ±3.03% (85 runs sampled)
EventEmitter3@0.1.6 x 2,496,186 ops/sec ±1.22% (81 runs sampled)
EventEmitter3(master) x 4,491,928 ops/sec ±2.01% (89 runs sampled)
fastemitter x 2,519,292 ops/sec ±0.97% (90 runs sampled)
event-emitter x 677,119 ops/sec ±1.75% (83 runs sampled)
contra/emitter x 106,522 ops/sec ±1.94% (88 runs sampled)
Fastest is EventEmitter3(master)

Starting benchmark context.js

EventEmitter1 x 1,002,678 ops/sec ±2.37% (85 runs sampled)
EventEmitter2 x 941,978 ops/sec ±1.82% (78 runs sampled)
EventEmitter3@0.1.6 x 9,530,584 ops/sec ±3.10% (86 runs sampled)
EventEmitter3(master) x 9,750,261 ops/sec ±3.56% (85 runs sampled)
Drip x 928,435 ops/sec ±2.69% (80 runs sampled)
fastemitter x 922,226 ops/sec ±1.27% (86 runs sampled)
event-emitter x 734,949 ops/sec ±1.22% (84 runs sampled)
contra/emitter x 96,574 ops/sec ±0.87% (84 runs sampled)
Fastest is EventEmitter3(master)
```